refactor(cli): move assertSupportedNodeVersion from shared/ to cli-core/
Previously: src/cli/src/shared/node-version.ts
Now:        src/cli/src/cli-core/node-version.ts

Rationale: assertSupportedNodeVersion() is a CLI startup validation — it
checks that the runtime Node.js version meets the minimum requirements
before any CLI work begins. This responsibility belongs in cli-core/
alongside the other infrastructure pieces (version.ts for plugin version,
errors.ts for error types, env-overrides.ts for env-based overrides)
rather than in shared/, which is intended for cross-command utilities.

The file was not exported through shared/index.ts, confirming it was
never genuinely "shared" — it was accessed via direct relative imports
from two commands (generate-gml-identifiers and generate-feather-metadata).
Moving it to cli-core/ makes it discoverable through the cli-core/index.ts
barrel and aligns the ownership model with the other startup checks.

- Update both command files to import from ../cli-core/node-version.js
- Re-export assertSupportedNodeVersion from cli-core/index.ts
- Add TSDoc to the function and a module-level placement rationale comment
- Add cli/test/core-node-version.test.ts following the core-*.test.ts convention

Agent-Logs-Url: https://github.com/SimulatorLife/GMLoop/sessions/2ab30a71-09e2-4bf1-9494-fc88bcf93059
