+++ src/cli/src/cli-core/index.ts
@@ -12,4 +12,5 @@ export * from "./errors.js";
 export * from "./extension-normalizer.js";
 export * from "./format-command-options.js";
 export * from "./main-module-runner.js";
+export * from "./node-version.js";
 export * from "./version.js";
